+++ app/CocoaPods/CPHomeWindowController.h
@@ -2,7 +2,13 @@
 
 @interface CPHomeWindowController : NSWindowController <NSDraggingDestination>
 
+/// Checks if binstab is already installed
+- (BOOL)isBinstubAlreadyInstalled;
+
 /// Forces the installation of the binstub
 - (IBAction)installBinstub:(id)sender;
+
+/// Removes alraedy installed binstub
+- (IBAction)removeBinstub:(id)sender;
 
 @end